coverter code split updated
