+++ tools/importer/transformers/image.js
@@ -1,0 +1,11 @@
+import {
+  imagetext,
+} from './util.js';
+
+const createImage = (main, document) => {
+  const imageText = main.querySelectorAll('div.imagetext');
+  [...imageText].forEach((imgText) => {
+    imagetext(imgText, document);
+  });
+};
+export default createImage;
