+++ tools/importer/transformers/pdfEmbed.js
@@ -1,0 +1,9 @@
+import { pdfembed } from './util.js';
+
+const createPDFEmbed = (main, document) => {
+  const pdfViewer = main.querySelectorAll('div.pdfviewer');
+  pdfViewer.forEach((pdf) => {
+    pdfembed(pdf, document);
+  });
+};
+export default createPDFEmbed;
